FastAPI/requirements_fastapi.txt
# FastAPI版 AI求人マッチングシステム 依存パッケージ

# FastAPI関連
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# セキュリティ
Werkzeug==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# データベース
psycopg2-binary==2.9.9
pgvector==0.2.0

# AI / 機械学習
openai==1.12.0
scikit-learn==1.3.2
numpy==1.26.2

# データ処理
pandas==2.1.4

# 環境変数
python-dotenv==1.0.0

# その他
uuid==1.30